+++ database/migrations/2021_03_05_200845_buat_pembelian_detail_table.php
@@ -17,9 +17,9 @@ class BuatPembelianDetailTable extends Migration
             $table->increments('id_pembelian_detail');
             $table->integer('id_pembelian');
             $table->integer('id_produk');
-            $table->integer('harga_beli');
-            $table->integer('jumlah');
-            $table->integer('subtotal');
+            $table->float('harga_beli');
+            $table->float('jumlah');
+            $table->float('subtotal');
             $table->timestamps();
         });
     }
