+++ database/migrations/2021_03_05_200853_buat_penjualan_table.php
@@ -17,10 +17,10 @@ class BuatPenjualanTable extends Migration
             $table->increments('id_penjualan');
             $table->integer('id_member');
             $table->integer('total_item');
-            $table->integer('total_harga');
+            $table->float('total_harga');
             $table->tinyInteger('diskon')->default(0);
-            $table->integer('bayar')->default(0);
-            $table->integer('diterima')->default(0);
+            $table->float('bayar')->default(0);
+            $table->float('diterima')->default(0);
             $table->integer('id_user');
             $table->timestamps();
         });
